feat(demo-integrations): move `goToDemoPage` inside `cy` commands
